[Playground CLI] Enable JavaScript to run the CLI by calling a runCLI function
