--- packages/playground/cli/jest.config.ts
@@ -1,13 +1,0 @@
-/* eslint-disable */
-export default {
-	displayName: 'nx-extensions',
-	preset: '../../../jest.preset.js',
-	transform: {
-		'^.+\\.[tj]s$': [
-			'ts-jest',
-			{ tsconfig: '<rootDir>/tsconfig.spec.json' },
-		],
-	},
-	moduleFileExtensions: ['ts', 'js', 'html'],
-	coverageDirectory: '../../../coverage/packages/nx-extensions',
-};
