+++ packages/playground/cli/src/server.ts
@@ -1,15 +1,19 @@
-import express, { Request } from 'express';
-import { PHPRequest, PHPResponse } from '@php-wasm/universal';
-import { IncomingMessage, Server, ServerResponse } from 'http';
-import { AddressInfo } from 'net';
+import type { PHPRequest, PHPResponse } from '@php-wasm/universal';
+import type { Request } from 'express';
+import express from 'express';
+import type { IncomingMessage, Server, ServerResponse } from 'http';
+import type { AddressInfo } from 'net';
+import type { RunCLIServer } from './run-cli';
 
 export interface ServerOptions {
 	port: number;
-	onBind: (port: number) => Promise<any>;
+	onBind: (server: Server, port: number) => Promise<RunCLIServer>;
 	handleRequest: (request: PHPRequest) => Promise<PHPResponse>;
 }
 
-export async function startServer(options: ServerOptions) {
+export async function startServer(
+	options: ServerOptions
+): Promise<RunCLIServer> {
 	const app = express();
 
 	const server = await new Promise<
@@ -42,7 +46,7 @@ export async function startServer(options: ServerOptions) {
 
 	const address = server.address();
 	const port = (address! as AddressInfo).port;
-	await options.onBind(port);
+	return await options.onBind(server, port);
 }
 
 const bufferRequestBody = async (req: Request): Promise<Uint8Array> =>
@@ -52,7 +56,7 @@ const bufferRequestBody = async (req: Request): Promise<Uint8Array> =>
 			body.push(chunk);
 		});
 		req.on('end', () => {
-			resolve(Buffer.concat(body));
+			resolve(new Uint8Array(Buffer.concat(body)));
 		});
 	});
